requirements.txt
openai>=1.30,<3.0
chromadb>=1.0,<2.0
gradio>=5.0,<6.0
pydantic>=2.7,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
pandas>=2.2,<3.1
openpyxl>=3.1,<4.0
python-pptx>=1.0,<2.0
python-docx>=1.1,<2.0
rank-bm25>=0.2,<0.3
rapidfuzz>=3.9,<4.0
tenacity>=8.2,<10.0
Pillow>=10.0,<12.0
orjson>=3.10,<4.0
PyYAML>=6.0,<7.0
